+++ unix/mclock.mli
@@ -18,6 +18,4 @@
 
     Clock returning monotonic time since an arbitrary point. To be
     used for eg. profiling. *)
-include Mirage_clock_lwt.MCLOCK
-
-val connect : unit -> t io
+include Mirage_clock.MCLOCK